backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
paramiko==3.4.0
Pillow>=10.2.0
aiofiles==23.2.1

